move relational mappers over to `full_batch_generators.py`
